refactor(security): centralize sanitizeImgUrl to lib/seo per CodeRabbit review
