+++ components/DynamicSearchIsland.tsx
@@ -19,41 +19,8 @@ import { getNextHighlightIndex } from '../services/suggestInteraction';
 import { buildPersonCardPresentation } from '../services/personPresentation';
 import { useDebounce } from '../hooks/useDebounce';
 import { apiGet } from '../lib/apiClient';
-import { safeImgUrl } from '../lib/seo';
+import { safeImgUrl, sanitizeImgUrl } from '../lib/seo';
 import '../styles/dynamic-search-island.css';
-
-// Local sanitizer to satisfy CodeQL taint tracker
-const sanitizeImgUrl = (url: string | null | undefined): string => {
-  const safe = safeImgUrl(url);
-  if (!safe) return '';
-  if (safe.startsWith('/') && !safe.startsWith('//')) return safe;
-  if (safe.startsWith('data:image/')) return safe;
-  try {
-    const parsed = new URL(safe);
-    if (parsed.protocol === 'http:' || parsed.protocol === 'https:') {
-      const host = parsed.hostname.toLowerCase();
-      if (
-        host === 'image.tmdb.org' ||
-        host === 'static.tvmaze.com' ||
-        host === 'images.unsplash.com' ||
-        host === 'graph.facebook.com' ||
-        host === 'avatars.githubusercontent.com' ||
-        host === 'moviemonk-ai.vercel.app' ||
-        host === 'googleusercontent.com' ||
-        host === 'lh3.googleusercontent.com' ||
-        host.endsWith('.googleusercontent.com') ||
-        host.endsWith('.supabase.co') ||
-        host === 'localhost' ||
-        host === '127.0.0.1'
-      ) {
-        return parsed.toString();
-      }
-    }
-  } catch (e) {
-    // ignore
-  }
-  return '';
-};
